pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
force-exclude = '[^/]*/migrations/[^/]*.py'

[tool.poetry]
name = "blab-controller"
version = "0.0.1dev"
description = "The back-end server upon which BLAB is built."
authors = ["C4AI"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
overrides = "^6.1.0"
channels = "^3.0.4"
channels-redis = "^3.3.1"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
django-cors-headers = "^3.11.0"
python-dotenv = "^0.19.2"
PyYAML = "^6.0"
uritemplate = "^4.1.1"
django-structlog = "^2.2.0"
celery = "^5.2.6"
redis = "^4.3.1"
drf-spectacular = { extras = ["sidecar"], version = "^0.25.1" }


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
darglint = "*"
flake8 = "*"
flake8-annotations = "*"
flake8-docstrings = "*"
flake8-bugbear = "*"
python-lsp-server = { extras = ["all"], version = "*" }
pyupgrade = "*"
mypy = "^0.931"
pre-commit = "^2.16.0"
black = "^22.3.0"
ruff = "^0.0.261"

[tool.poetry.group.PostgreSQL]
optional = true
[tool.poetry.group.PostgreSQL.dependencies]
psycopg2-binary = "^2.9.3"

[tool.poetry.group.MySQL]
optional = true
[tool.poetry.group.MySQL.dependencies]
mysqlclient = "^2.1.0"

[tool.poetry.group.Oracle]
optional = true
[tool.poetry.group.Oracle.dependencies]
cx-Oracle = "^8.3.0"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.0.1"
sphinx-autoapi = "^2.1.0"
furo = "^2023.5.20"

[tool.ruff]
select = ["F", "E", "W", "C90", "I", "N", "D", "UP", "YTT", "ANN", "S", "BLE", "B", "A", "C4", "T10", "DJ", "EM", "EXE", "ISC", "ICN", "ISC002", "ISC003", "ICN", "G", "INP", "PIE", "PYI", "PT", "Q", "RSE", "RET", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["ANN002", "ANN003", "ANN1", "ANN204", "ANN401", "ARG002", "A003", "D105", "D106", "D203", "D213", "D417", "DTZ", "I001", "TRY00"]
exclude = ["blab_controller/chat/migrations"]
per-file-ignores = { "blab_controller/controller/settings/*.py" = ["F401", "F403", "F405"], "test*.py" = ["D", "S101"], "*_TEMPLATE.py" = ["N999"], "manage.py" = ["INP001"] }

[tool.ruff.pydocstyle]
ignore-decorators = ["overrides.overrides"]
